hack-bitcoin-bitcoin-django-backend-main/pyproject.toml
[tool.poetry]
name = "bitcoin_django_backend"
version = "0.0.1"
description = "bitcoin django backend"
authors = ["O Boticario <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9.7"
django = "^4.0"
django-cors-headers = "^3.13.0"
django-filter = "^21.1"
django-localflavor = "^3.0"
django-structlog = "^2.1.0"
djangorestframework = "^3.11.0"
djangorestframework-camel-case = "^1.2.0"
drf-spectacular = "^0.22.1"
gunicorn = "^20.0.4"
uvicorn = "^0.17.6"
requests = "^2.23.0"
whitenoise = "^6.2.0"
dj-database-url = "^0.5.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.4"
black = "^22.3"
flake8 = "^4.0.1"
ipython = "^8.4.0"
isort = "^5.2.2"
model_bakery = "^1.5.0"
mypy = "^0.961"
pdbpp = "^0.10.2"
pytest = "^6.0.1"
pytest-cov = "^3.0.0"
pytest-deadfixtures = "^2.1.0"
pytest-django = "^4.1.0"
pytest-env = "^0.6.2"
pytest-pythonpath = "^0.7.4"
pytest-vcr = "^1.0.2"
safety = "^1.8.5"
vcrpy = "^4.1.0"

[tool.black]
line-length = 79
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )
)
'''

[tool.isort]
line_length = 79
known_first_party = "bitcoin_django_backend"
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true
skip_glob = [
    "**migrations**"
]

[tool.coverage.run]
omit = [
    'src/manage.py',
    '**asgi.py',
    '**wsgi.py',
    '**migrations**',
    '**tests**',
    '**settings**',
]
branch = true

[build-system]
requires = ["poetry>=1.1.12"]
build-backend = "poetry.masonry.api"